Remove references to type: :new test quarantine tag
We decided to remove the `type: :new` quarantine type for tests per
https://gitlab.com/gitlab-com/www-gitlab-com/-/merge_requests/47681 .
New tests will be unquarantined by default and only moved to quarantine
if they fail (and the fix cannot be made quickly).  Once the new tests have
passed outside of quarantine for 7 consecutive days, they should then be promoted to `:reliable`.