Automatically draft release post items
Issues contain sufficient information to draft a release post item, but
using command line tools to do this it not efficient or easy for
everyone. This commit implements API support and a CI job for automating
the process.

When an issue is ready for a release post item to be drafted, add a
release post item label (e.g. 'release post item:secondary'). The script
will:

- find all issues with release post item labels the `gitlab-org` group
- create draft release post items
- commit them to unique branches
- create a merge request and assign it to the group product manager

The script can be run manually using the `--no-local` option to use the
GitLab API for all read/write operations.

The script can be used for creating a release post item for a single
issue by providing the issue URL as an argument when running the
command.