Introduce error budgets for Runner team
Recently the Runner team has had a few large regressions such as
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/26609 and
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/26347. This is
normal since we rapidly increased the head count of the team and our
current process that we used when we were just 2 people doesn't scale.
We need to find a balance between adding new features and making sure
that we are keep reliability as a priority.