Only evaluate 'Release notes' section
Only the section titled 'Release notes' (case insensitive) is used for
determining the release post entry description, documentation_link and
image_url.

The image_url will be taken as the first image in the release notes
section.

The documentation_link will be taken as the first URL in the release
notes section that begins with 'https://docs.gitlab.com'