PagerDuty Overrides made by covering team member
This change explicitly sets the responsibility of configuring the
override in PagerDuty with the team member who is volunteering, or
being told, that they will provide coverage for a team member that
has requested someone take their shift for an absence&mdash;
planned or unplanned.

This change ensures that someone is not placed on-call without
their explicit knowledge of the change.