Add new hire shadowing responsibility
We have a lot of pipelines and a lot of tools that we use during the pipeline
on-call rotation.  The on-call process can be daunting for first-time DRIs.  In order to help them become familiar with our process and tools more quickly, the
new hire should spend at least a few days shadowing and being shadowed by more
experienced members of the department.