Update incident management workflow
The current implementation of the Slack `/incident` command is soon to
be replaced by Woodhouse:
https://gitlab.com/gitlab-com/gl-infra/woodhouse.

The previous documentation stated that the engineer on-call would be
notified, but this isn't necessarily the case as they were not paged.
Woodhouse optionally allows the engineer to be paged. Emphasise this
choice.

A review of Woodhouse's incident management functionality from its
staging deployment can be found here:
https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/11573

Issue to start using Woodhouse for incident management:
https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/11602