calibration: Require notes 12 hours ahead of time
When notes on ICs are added minutes before the meeting, it removes the
option to better calibrate your own ICs, as data points on others aren't
available. By requiring notes 12 hours ahead of time the EM could still
update and iterate on their own notes, and it also allows for better
collaboration because others can feed of these notes too.