Add Engineering IC Leadership page
This is part of the [IC Gearing WG][WG].

Here we introduce an Engineering IC Leadership page explaining Staff+
roles, it does not replace the job role definitions, but it provide a
common framework to explain those roles.

The current page does not refer to any new role (like the Principal role
proposed in
https://gitlab.com/gitlab-com/www-gitlab-com/-/merge_requests/59465 )
and could be merged as a description of the status quo.

Nevertheless it will be easier to better define Staff+ roles, as well as
a new role between Staff and Distinguished, with the help of this new
page.

[WG]: https://about.gitlab.com/company/team/structure/working-groups/ic-gearing/