Add Reimbursable Amount column to hardware table
While the average price column is helpful to give
people a rough sense of the price range for the
hardware, it is not clear that it is also the
maximum reimbursable amount.

A Reimbursable Amount column is introduced
containing the higher amount from the average
price column to make this clearer.