Fix PDF-Diagram links opening within object
Previously if you clicked a link in the PDF-Diagram in the context
of the documentation, the link would open inside the HTML object on
the page, instead of navigating the entire page to that anchor. This
fixes that problem by adding target attributes to the links which
target the parent element of the object - the page itself.