Revise working-by-default product principle
The distinction between on-by-default and setup-by-default is important.
The previous definitions did not clearly address situations where extra
components are required. The distinction is revised here based on:

- setup tasks can be completed by Owners, Maintainers and Developers
- install tasks can only be completed by Admins and those with access to
  the host machine

On-by-default, which was previously defined primarily in terms of
feature flags, maps to the second situation because feature flags can
only be toggled by Admins using the API or people with root access via
the Ruby console. On-by-default has now been extended to include other
tasks like configuration via gitlab.rb, or installing extra components.

Setup-by-default, now calls for thoughtful migrations to enable new
features and great defaults for existing users/groups/projects. This was
previously described in the on-by-default section despite relating to
non-admin existing users not needing to perform any setup for new
features.