De-duplicate content in Spending Company Money
The Expenses page is considered the "single source of truth" for
information about expenses and reimbursement. The Spending Company
Money page duplicated a large amount of content from the Expenses
page, instead of simply linking to that page. This removes the
duplicate content. This also makes a few small updates to the Expenses
page, to reflect changes that were only made on the Spending Company
Money page (illustrating the problem with duplicate content).