Add soft cutoff date for Dev:Manage MRs
The Manage sub-department has experienced challenges with the
release process for code merged late in the release window (near
the 17th). This introduces a soft cutoff date to the Manage handbook
page, as a guideline to help make decisions about what will be
considered part of a release and hopefully reduce stress around the
release process.