Warn against release posts on the 17th
Code merged on the 17th is not guaranteed to be part of a release,
so release posts should probably not be merged for code merged on
the 17th. Also, discouraging release posts based on last-minute code
merges will help alleviate some of the urgency and stress centered
on the 17th.