Contribute to the retro issue weekly
We have found that we forget what happened throughout
the milestone if we wait for too long to add items to the retro issue.
To help with that, we are adding a small adendum to our retro
section to show that we want to add items to the issue throughout
the milestone instead of waiting for the end.