Remove start date from files + where it  is used
This disables the hiring charts as it is 100% based on the
start date in the files. This data should be in Sisense and
we can work with the data team to recreate them. However a
leadership decision should be made to see if we want to have these
in the public handbook.