Uncouple the Distribution Triage and Demo duties
The Distribution team had traditionally used the triage rotation as a
way to select a weekly demo presenter. This rarely works out as the team
member on triage is not necessarily also the person with a project ready
to demo in a particular week.  This merge request decouples the two
tasks and reflects the actual current process.

Signed-off-by: Robert Marshall <rmarshall@gitlab.com>