Update terminology to use Incident Manager instead of IMOC
This MR shifts our terminology to use Incident Manager instead IMOC in
most contexts.

Incident Manager On Call (IMOC) does not make sense in many contexts as
it conflates the role of Incident Manager with on-call responsibilities.
For example, it is common to have actively engaged Incident Managers who
are not on call.