Clarify that hiring nice-to-haves are role-specific
New hires are required to satisfy 5 of 9 "nice to have" requirements.
However, there could be confusion about what those requirements are,
because the list that appears in Greenhouse scorecards is not the
same as the role-specific lists that appear in the Job Families
documentation. This modifies the hiring documentation to clarify that
we should use the role-specific lists (and adds a a link to them).