Update the Configure planning doc
Requires that UX, backend, and or Frontend labels are added during the
breakdown fase.

This will give more clear visibility when picking up the issues that are
read_for_development.

Also, it helps us to more easily define if more people need to be
mentioned/assigned to a certain issue.