Be more explicit about FRT and SLA abbreviation
Neither of those abbreviations are explicitely spelled out anywhere else
and both are used very ubiquitisly within the support team.

"Support Engineer Responsibilities" and "Working on Tickets" both are
early on in the Support Engineer learning path, so adding it here should
help most people see it soon enough.
Additionally, "Meeting the FRT Service Level Agreement (SLA)" is the top
search result for "FRT" in the handbook and should have it spelled out
once as well.