Updates to the base theme
All layouts with the exception of the one supporting the homepage have
been moved to the theme so they can be shared across handbooks.  The
same applies to the contribuitor docs.

The handbook has been tweaked to remove references to legacy handbook to
current public handbook to avoid any confusion and the background on the
top cover has been altered slightly.