Final fixes for handbook migration script
These should be the final fixes for the handbook migration script before
its used to migrate the job-families site to the new handbook.  Known
issues are:

* Creates duplicate target entries in the redirect.yml
* Markdownlint can't fix all errors
* Doesn't create all required _index.md entries for all job families

This script does the hard part of moving all the content files and
preserving their histories but to keep the script pretty generic we stop
there.