Add company section to handbook and update script
Updated the script to support moving indvidual parts of company and
engineering sections of the Handbook.  Added a new company section to
the handbook as Company doesn't have an index page of its own.  Also
added an empty engineering directory so we can start to migrate
engineering to the Handbook.