Added vision back in to the Handbook
The vision pages are controlled by Buyer expereicne and as such lives on
the about site but its valulable content for the Handbook as well.  To
that end the included shortcode gets the content from Buyer Expereicne
and builds a standard page based on the markdown in the Yaml file.

We stick to a single source of truth being the version in Buyer
experience.  We'll chase any breaking changes they make to the yaml
file.