Bring the vision content back to the handbook
Following agreement with Lauren from buyer experience I've recreated the
vision page in the new handbook site.  As the original page was deleted
I'm unable to get the history for this page.

I've also added a new include to retain the sales links at the bottom of
the page to point back to the free trial option and about.gitlab.com/sales.