Update CODEOWNERS to add egroup group
Its been agreed all mandated entries in the CODEOWNERS file should have
an egroup entry.  E-Group are the ultimate owners of all handbook
content and be reviewers and approvers of last resort for Handbook
content and the ultimate DRIs.

Additionally we've added a handbook group to all entries as there are
activities which sometimes make mass changes to the handbook which can
touch on many pages across many sections.  This group allows the
Handbook team to act without having to gather many approvals.