Update Codeowners, add operations template
This is a mix of items in this MR.  First update the CODEOWNERs file
moving required entries to the bottom as these need to come last.

Relocated my own README as I am no longer a member of the CoST team and
my README currently is without a sensible home.

Added a handbook-operations template which automatically includes the
commit messages from the MR.  Also updates the label to
handbook::operations from handbook::content.

Finally moves the `content/_index.md` file to `content/_index.html` to
be consistent with the internal-handbook and help clean up that
codeowner rule.