Prepare structure for Observability team
The current information about Scalability groups is outdated after the
restructure last year. While we're currently working out direction and
vision and more content is expected to come, I'd like to prepare a new
team page and structure for the Observability group.

I expect we can make use of this going forward to discuss team
agreements.