Implement improved controlled document handling
This rather large MR improves controlled document handling and
implements linting against the CODEOWNERS file and Controlled Documents.
What this means is that we'll check for missing files listed in
CODEOWNERS and duplicate entries.  We'll also check to make sure files
listed in the `[controlled-documents]` section of CODEOWNERS are listed
as such in the files front matter and that any file who's front matter
says its a controlled_document that it has a matching entry in
CODEOWNERS.